Fix hàm check đầu vô cho tất cả api để chuẩn bị việc sửa api, gần xong phrase 4, sửa nhỏ mà nhiều

+++ StoreManagement/Controllers/Class/Client_DemandDetail.cs
@@ -28,7 +28,7 @@ namespace StoreManagement.Class
 
         public int DemandedItemPK { get; set; }
 
-        public int DemandedQuantity { get; set; }
+        public double DemandedQuantity { get; set; }
 
         public string AccessoryID { get; set; }
